AI_CRM_FULL_ASSIGNMENT/backend/requirements.txt

fastapi
uvicorn
langgraph
langchain
groq
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
